requirements.txt
pytest==7.1.2
selenium==4.1.3
allure-pytest==2.9.45
python-dotenv==0.20.0
